Removed useless abstract logging. There is no need in such abstraction, also dynamic resolving works bad with async tasks

+++ src/L2dotNET.Models/Contracts/SpawnlistContract.cs
@@ -23,6 +23,6 @@ namespace L2dotNET.DataContracts
 
         public int RespawnRand { get; set; }
 
-        public int PeriodOfDay { get; set; }
+        public byte PeriodOfDay { get; set; }
     }
 }
